strings revisited

+++ src/Compiler/AbiLowering.cs
@@ -94,7 +94,7 @@ internal sealed class AbiLowerer
         return type.Kind switch
         {
             StarkTypeKind.Ascii => StarkTypeSymbols.RawPointer(StarkTypeSymbols.Integer(8), isMutable: false),
-            StarkTypeKind.Unicode when !forReturnValue => StarkTypeSymbols.RawPointer(StarkTypeSymbols.Integer(8), isMutable: false),
+            StarkTypeKind.Unicode when !forReturnValue => StarkTypeSymbols.RawPointer(StarkTypeSymbols.Integer(32), isMutable: false),
             _ => type
         };
     }
